requirements.txt
numpy
pandas
gurobipy
geopandas
pyyaml
ortools
matplotlib
descartes
seaborn